Factor out project.util.phone_number. Fixes #457.
